requirements.txt
Flask==1.1.1
werkzeug==1.0.1
numpy==1.18.1
tensorflow==2.3.1
efficientnet==1.1.1
Pillow==7.0.0
keras==2.4.3
